вынесли логику работы со state из фрагмента по поиску вакансий в SearchFragmentStateHandler
