+++ app/src/main/java/ru/practicum/android/diploma/search/ui/SearchFragment.kt
@@ -6,8 +6,6 @@ import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
@@ -19,11 +17,9 @@ import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.databinding.FragmentSearchBinding
 import ru.practicum.android.diploma.job.ui.JobFragment
 import ru.practicum.android.diploma.search.domain.model.Vacancy
-import ru.practicum.android.diploma.search.presentation.SearchScreenState
 import ru.practicum.android.diploma.search.presentation.SearchViewModel
 import ru.practicum.android.diploma.search.ui.adapter.SearchAdapter
 import ru.practicum.android.diploma.search.ui.adapter.SearchClickListener
-import ru.practicum.android.diploma.util.Formatter
 
 class SearchFragment : Fragment(), SearchClickListener {
 
@@ -31,24 +27,25 @@ class SearchFragment : Fragment(), SearchClickListener {
     private val binding get() = _binding!!
     private val viewModel by viewModel<SearchViewModel>()
     private var textWatcher: TextWatcher? = null
-    private var searchAdapter = SearchAdapter(this)
+    private val searchAdapter = SearchAdapter(this)
     private var clickDebounce: ((Boolean) -> Unit)? = null
     private var isClickAllowed = true
+    private var stateHandler: SearchFragmentStateHandler? = null
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         _binding = FragmentSearchBinding.inflate(inflater, container, false)
-
+        stateHandler = SearchFragmentStateHandler(binding, searchAdapter, requireContext())
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         viewModel.getScreenState().observe(viewLifecycleOwner) { state ->
-            renderState(state)
+            stateHandler?.renderState(state)
         }
 
         binding.apply {
@@ -119,134 +116,7 @@ class SearchFragment : Fragment(), SearchClickListener {
         textWatcher.let { binding.searchField.removeTextChangedListener(it) }
         _binding = null
         clickDebounce = null
-    }
-
-    private fun renderState(state: SearchScreenState) {
-        when (state) {
-            is SearchScreenState.Default -> showDefaultScreenState()
-            is SearchScreenState.Loading -> showProgressbar()
-            is SearchScreenState.InternetConnectionError -> showInternetConnectionError()
-            is SearchScreenState.ServerError -> showServerError()
-            is SearchScreenState.SearchError -> showSearchError()
-            is SearchScreenState.ShowContent -> showContent(state.vacancies, state.found)
-            is SearchScreenState.UploadNextPage -> uploadNextPage()
-            is SearchScreenState.Error -> showError()
-            is SearchScreenState.IOError -> showIOError()
-        }
-    }
-
-    private fun showProgressbar() {
-        binding.apply {
-            searchScreenCover.isVisible = false
-            progressBar.isVisible = true
-            errorPlaceholder.isVisible = false
-            searchStatus.isVisible = false
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun showInternetConnectionError() {
-        binding.apply {
-            searchScreenCover.isVisible = false
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = true
-            errorPlaceholder.setText(R.string.no_internet)
-            errorPlaceholder.setCompoundDrawablesWithIntrinsicBounds(
-                0,
-                R.drawable.img_internet_connection_error,
-                0,
-                0
-            )
-            searchStatus.isVisible = false
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun showServerError() {
-        binding.apply {
-            searchScreenCover.isVisible = false
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = true
-            errorPlaceholder.setText(R.string.server_error)
-            errorPlaceholder.setCompoundDrawablesWithIntrinsicBounds(
-                0,
-                R.drawable.img_server_error_placeholder,
-                0,
-                0
-            )
-            searchStatus.isVisible = false
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun showSearchError() {
-        binding.apply {
-            searchScreenCover.isVisible = false
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = true
-            errorPlaceholder.setText(R.string.empty_search_results)
-            errorPlaceholder.setCompoundDrawablesWithIntrinsicBounds(
-                0,
-                R.drawable.img_empty_search_results,
-                0,
-                0
-            )
-            searchStatus.isVisible = true
-            searchStatus.text = getString(R.string.no_such_vacancies)
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun showDefaultScreenState() {
-        binding.apply {
-            searchScreenCover.isVisible = true
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = false
-            searchStatus.isVisible = false
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun uploadNextPage() {
-        binding.apply {
-            errorPlaceholder.isVisible = false
-            progressBar.isVisible = true
-        }
-    }
-
-    private fun showError() {
-        Toast.makeText(requireContext(), getString(R.string.error_occured), Toast.LENGTH_SHORT).show()
-        binding.apply {
-            searchScreenCover.isVisible = true
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = false
-            searchStatus.isVisible = false
-            recyclerView.isVisible = false
-        }
-    }
-
-    private fun showIOError() {
-        Toast.makeText(requireContext(), getString(R.string.error_occured), Toast.LENGTH_SHORT).show()
-        binding.apply {
-            searchScreenCover.isVisible = false
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = false
-            searchStatus.isVisible = true
-            recyclerView.isVisible = true
-        }
-    }
-
-    private fun showContent(list: ArrayList<Vacancy>, resultsQty: Int) {
-        searchAdapter.vacanciesList.clear()
-        searchAdapter.vacanciesList.addAll(list)
-        searchAdapter.notifyDataSetChanged()
-        binding.apply {
-            progressBar.isVisible = false
-            errorPlaceholder.isVisible = false
-            searchStatus.isVisible = true
-            searchStatus.text = getVacanciesWordForm(resultsQty)
-            recyclerView.isVisible = true
-        }
+        stateHandler = null
     }
 
     override fun onVacancyClick(vacancy: Vacancy) {
@@ -262,26 +132,6 @@ class SearchFragment : Fragment(), SearchClickListener {
                 vacancyId = vacancyId
             )
         )
-    }
-
-    private fun getVacanciesWordForm(quantity: Int): String {
-        val vacancyForm = Formatter
-            .quantityWordFormFormatter(
-                quantity,
-                getString(R.string.vacancy),
-                getString(R.string.vacancy_genitive),
-                getString(R.string.vacancies_genitive)
-            )
-
-        val foundForm = Formatter
-            .quantityWordFormFormatter(
-                quantity,
-                getString(R.string.found_f),
-                getString(R.string.found_f_genitive),
-                getString(R.string.found_f_genitive_pl)
-            )
-
-        return "$foundForm $quantity $vacancyForm"
     }
 
     private fun allowClick() {
